-melfas_ts.c - only log touchscreen events when debug mode is on. This just puts a lot of junk into the kernel log that make it hard to troubleshoot other things.
-Removing more CONFIG_ICS conditionals. Since this version of the kernel
will be ICS only, I am removing the complexities. I am removing in
batches so that testing is easier to confirm issues on removal. atmel is
not used for p5, but it was cleaned up too to make searching for the
tags easy for the next batch to be removed.