revert - board-p5 (p5 and p5wifi): uartd's parent to pll_p and pll_a to 56448000
Backing this change out for now since some are reporting a lot more
random reboots than in the prior build, and this was one of the
unknowns.

We received official HC 3.2 source from Samsung recently for the 7310
and this was still the same as it was in the 3.1 source. We have lived
with it this long and will re-implement later if more documentation
surfaces that proves it is the right fix.